+++ src/plotlines.rs
@@ -52,8 +52,8 @@ impl<'p> PlotLines<'p> {
     }
 
     #[inline]
-    pub fn graph_size(mut self, graph_size: ImVec2) -> Self {
-        self.graph_size = graph_size;
+    pub fn graph_size<S: Into<ImVec2>>(mut self, graph_size: S) -> Self {
+        self.graph_size = graph_size.into();
         self
     }